+++ Tests/AzureFunctions.TestApplication/AzureFunctions.TestApplication.Api/MappedAzureFunction.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Net;
+using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Transactions;
@@ -18,7 +19,6 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
 using Microsoft.OpenApi.Models;
-using Newtonsoft.Json;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.AzureFunctions.AzureFunctionClass", Version = "1.0")]
@@ -50,7 +50,7 @@ namespace AzureFunctions.TestApplication.Api
             try
             {
                 var requestBody = await new StreamReader(req.Body).ReadToEndAsync();
-                var request = JsonConvert.DeserializeObject<SampleMappedRequest>(requestBody)!;
+                var request = JsonSerializer.Deserialize<SampleMappedRequest>(requestBody)!;
                 await _validator.Handle(request, cancellationToken);
 
                 using (var transaction = new TransactionScope(TransactionScopeOption.Required,