+++ Modules/Intent.Modules.AzureFunctions/release-notes.md
@@ -3,6 +3,7 @@
 - Update: Improved Queue Trigger support, including receiving messages as `QueueMessage` and output binding support for `Azure Functions`.
 - Update: Removed compiler warning.
 - Update: Services that produce Azure Functions will no longer dump its own functions in the root folder of the API project but in their own folders.
+- Update: Moved away from Newtonsoft JSON and now using System.Text.Json instead.
 
 ### Version 4.0.5